+++ src/werkzeug/formparser.py
@@ -63,7 +63,8 @@ def parse_form_data(
 ) -> t_parse_result:
     """Parse the form data in the environ and return it as tuple in the form
     ``(stream, form, files)``.  You should only call this method if the
-    transport method is `POST`, `PUT`, or `PATCH`.
+    transport method carries a body, such as `POST`, `PUT`, `PATCH`, or
+    `QUERY`.
 
     If the mimetype of the data transmitted is `multipart/form-data` the
     files multidict will be filled with `FileStorage` objects.  If the